requirements.txt
praw>=7.7.0
pandas>=2.0.0
python-dotenv>=1.0.0
